# Use `esbuild` to transpile frontend assets.